Resolvendo sextoBug. Corrigindo fórmula dos pães para retornar double.

+++ desafio-aula-1/registradora-cheia-de-bugs/src/RelacaoPesoPreco.java
@@ -1,6 +1,6 @@
 public class RelacaoPesoPreco {
 
-    public static double retornaPrecoProduto(String item, int qtd) {
+    public static double retornaPrecoProduto(String item, double qtd) {
         double precoTotal = 0;
 
         if ("pao".equals(item)) {
